+++ bootstrapping/build.netfx.ps1
@@ -17,7 +17,7 @@ $SolutionDirectory = "$PSScriptRoot\_SOLUTION_DIRECTORY_"
 $BuildProjectFile = "$PSScriptRoot\_BUILD_DIRECTORY_NAME_\_BUILD_PROJECT_NAME_.csproj"
 $BuildExeFile = "$PSScriptRoot\_BUILD_DIRECTORY_NAME_\bin\debug\_BUILD_PROJECT_NAME_.exe"
 
-$TempDirectory = "$PSScriptRoot\.tmp"
+$TempDirectory = "$PSScriptRoot\_ROOT_DIRECTORY_\.tmp"
 
 $NuGetUrl = "https://dist.nuget.org/win-x86-commandline/$NuGetVersion/nuget.exe"
 $NuGetFile = "$TempDirectory\nuget.exe"